refactor: modularize ConfigurationTab into focused section components
Breaks down the 1201-line ConfigurationTab component into 6 focused,
reusable section components, reducing the main file to 548 lines
(54% reduction).

**New structure:**
- src/components/configuration/types.ts - Shared TypeScript interfaces
- src/components/configuration/constants.ts - Role/preset/region data
- src/components/configuration/NodeIdentitySection.tsx - Node names
- src/components/configuration/DeviceConfigSection.tsx - Device role & settings
- src/components/configuration/LoRaConfigSection.tsx - Radio configuration
- src/components/configuration/PositionConfigSection.tsx - GPS/position settings
- src/components/configuration/MQTTConfigSection.tsx - MQTT module config
- src/components/configuration/NeighborInfoSection.tsx - NeighborInfo module

**Benefits:**
- ConfigurationTab: 1201 → 548 lines (54% reduction)
- Each section component < 300 lines
- Better testability and maintainability
- Eliminates duplicate constant arrays
- Improves code organization and separation of concerns

**Part of Phase 1 of the refactoring plan from docs/REFACTORING_PLAN.md**

✅ TypeScript compiles without errors
✅ All 293 tests passing
