+++ src/components/configuration/types.ts
@@ -1,16 +1,58 @@
-export interface DeviceConfig {
+/**
+ * Shared types for ConfigurationTab components
+ */
+
+export interface RoleOption {
+  value: number;
+  name: string;
+  shortDesc: string;
+  description: string;
+}
+
+export interface ModemPresetOption {
+  value: number;
+  name: string;
+  description: string;
+  params: string;
+}
+
+export interface RegionOption {
+  value: number;
+  label: string;
+}
+
+export interface ConfigSectionProps {
+  isSaving: boolean;
+  onSave: () => Promise<void>;
+}
+
+export interface NodeIdentityData {
   longName: string;
   shortName: string;
+}
+
+export interface DeviceConfigData {
   role: number;
   nodeInfoBroadcastSecs: number;
+}
+
+export interface LoRaConfigData {
+  usePreset: boolean;
   modemPreset: number;
   region: number;
   hopLimit: number;
+}
+
+export interface PositionConfigData {
+  positionBroadcastSecs: number;
+  positionSmartEnabled: boolean;
+  fixedPosition: boolean;
   fixedLatitude: number;
   fixedLongitude: number;
   fixedAltitude: number;
-  positionBroadcastSecs: number;
-  positionBroadcastSmartEnabled: boolean;
+}
+
+export interface MQTTConfigData {
   mqttEnabled: boolean;
   mqttAddress: string;
   mqttUsername: string;
@@ -20,7 +62,7 @@ export interface DeviceConfig {
   mqttRoot: string;
 }
 
-export interface ConfigurationSectionProps {
-  onSave?: () => void;
-  onError?: (error: string) => void;
+export interface NeighborInfoConfigData {
+  neighborInfoEnabled: boolean;
+  neighborInfoInterval: number;
 }
